Refactor csvparser into utils
